move projects back to app and fade in on load
